Harden readers against malformed input found by cargo-fuzz
Three fuzzing rounds over all 131 supported extensions turned up 34
distinct failures; every one now returns instead of aborting.

- EXIF: cap IFD nesting at 32 and never walk the same IFD twice, which
  overflowed the stack on self-referencing SubIFD/ExifIFD pointers.
- Lytro JSON: step over non-values in arrays (infinite loop) and cap
  nesting at 64 (quadratic tag paths).
- JPEG extended XMP: reserve the bytes actually received, not the
  declared total (a 3 GB allocation from a 1 KB file).
- FlashPix: sector shifts checked, sector chains bounded by file size
  (cyclic FATs copied up to 655 MB per stream).
- TNEF: stop multi-value loops that stop consuming data.
- Sony SR2: check a value run fits before reserving its declared count.
- Slices cut inside a UTF-8 character: vCard, HTML, GeoTIFF ASCII
  params, composite SubSec dates, RIFF, GoPro GPSU, SVG rdf:RDF range.
- Arithmetic overflow: PICT and PCX dimensions (now signed, as in
  ExifTool), TNEF FILETIME, MXF BER lengths, Google HDRP bit reader.
- LNK: item shorter than its type byte.

Adds the fuzz/ harness (first byte picks the extension) and one
regression input per format and failure kind under tests/malformed/.
Parity audit: 196/196 files identical, write cases ISO.

+++ src/formats/pict.rs
@@ -30,8 +30,8 @@ pub fn read_pict(data: &[u8]) -> Result<Vec<Tag>> {
     // Version 2 opcode: 0x0011 at bytes 10-11
     let mut h_res: Option<f64> = None;
     let mut v_res: Option<f64> = None;
-    let mut w = (right - left) as i32;
-    let mut h = (bottom - top) as i32;
+    let mut w = right as i32 - left as i32;
+    let mut h = bottom as i32 - top as i32;
 
     if d.len() >= 40 && d[10] == 0x00 && d[11] == 0x11 {
         // Version 2: next 2 bytes are 0x02ff, then check for extended
